Reealizo operación PUT en centro, con esto en Centro esta completo el CRUD, reorganizo y completa adaptación a MVP de centro
