+++ app/src/main/java/com/alberto/gesresfamilyapp/contract/centro/CentrosListContract.java
@@ -1,4 +1,4 @@
-package com.alberto.gesresfamilyapp.contract;
+package com.alberto.gesresfamilyapp.contract.centro;
 
 import com.alberto.gesresfamilyapp.domain.Centro;
