+++ app/src/main/java/com/alberto/gesresfamilyapp/view/residente/ResidentesListView.java
@@ -1,9 +1,4 @@
-package com.alberto.gesresfamilyapp.view;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
+package com.alberto.gesresfamilyapp.view.residente;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -11,30 +6,34 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 
-import com.alberto.gesresfamilyapp.R;
-import com.alberto.gesresfamilyapp.adapter.CentroAdapter;
-import com.alberto.gesresfamilyapp.contract.CentrosListContract;
-import com.alberto.gesresfamilyapp.domain.Centro;
-import com.alberto.gesresfamilyapp.presenter.CentrosListPresenter;
-import com.google.android.material.appbar.MaterialToolbar;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
+import com.alberto.gesresfamilyapp.R;
+import com.alberto.gesresfamilyapp.adapter.ResidenteAdapter;
+import com.alberto.gesresfamilyapp.contract.residente.ResidentesListContract;
+import com.alberto.gesresfamilyapp.domain.Residente;
+import com.alberto.gesresfamilyapp.presenter.residente.ResidentesListPresenter;
+import com.google.android.material.appbar.MaterialToolbar;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class CentrosListView extends AppCompatActivity implements CentrosListContract.view {
+public class ResidentesListView extends AppCompatActivity implements ResidentesListContract.view {
 
-    public static List<Centro> centroList = new ArrayList<>();
-    private CentroAdapter adapter;
+    public static List<Residente> residentesList = new ArrayList<>();
+    private ResidenteAdapter adapter;
 
-    private CentrosListPresenter presenter;
+    private ResidentesListPresenter presenter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_centros);
+        setContentView(R.layout.activity_residentes);
 
-        presenter = new CentrosListPresenter(this);
+        presenter = new ResidentesListPresenter(this);
 
         MaterialToolbar topAppBar = findViewById(R.id.topAppBar);
         setSupportActionBar(topAppBar);
@@ -51,8 +50,9 @@ public class CentrosListView extends AppCompatActivity implements CentrosListCon
     }
 
     private void initializeRecyclerView(){
-        centroList = new ArrayList<>();
-        RecyclerView recyclerView = findViewById(R.id.centros_list);
+        residentesList = new ArrayList<>();
+
+        RecyclerView recyclerView = findViewById(R.id.residentes_list);
         //esto le dice que tenga un tamaño fijo y que ocupe el máximo espacio asignado
         recyclerView.setHasFixedSize(true);
         //Esto le dice que lo va a gestionar un linear layout manager
@@ -60,19 +60,19 @@ public class CentrosListView extends AppCompatActivity implements CentrosListCon
         //así se ciñe al Layout manager
         recyclerView.setLayoutManager(layoutManager);
         //hago mi adapter propio no utilizo el arrayadapter de android
-        adapter = new CentroAdapter(this, centroList);
+        adapter = new ResidenteAdapter(this, residentesList);
         recyclerView.setAdapter(adapter);
     }
 
     @Override
     protected void onResume() {
         super.onResume();
-        presenter.loadAllCentros();
+        presenter.loadAllResidentes();
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.menu_options, menu);
+        getMenuInflater().inflate(R.menu.menu_options_add, menu);
         return true;
     }
 
@@ -80,15 +80,11 @@ public class CentrosListView extends AppCompatActivity implements CentrosListCon
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
         if (item.getItemId() == R.id.registrar) {
             //Con Intent de digo donde estoy y a donde quiero ir
-            Intent intent = new Intent(this, RegisterCentroView.class);
-            startActivity(intent);
-            return true;
-        } else if (item.getItemId() == R.id.view_map) { //Para cuando pulsan en la boton del mapa en el menu options
-            Intent intent = new Intent(this, MapsActivity.class); //donde nos manda al pinchar sobre el boton mapas en el menu options
+            Intent intent = new Intent(this, RegisterResidenteActivity.class);
             startActivity(intent);
             return true;
         }
-        return super.onOptionsItemSelected(item); // Llamar al método de la superclase
+        return false;
     }
 
     public void cancel(View view) {
@@ -96,12 +92,11 @@ public class CentrosListView extends AppCompatActivity implements CentrosListCon
     }
 
     @Override
-    public void showCentros(List<Centro> centros) {
-        centroList.clear();
-        centroList.addAll(centros);
+    public void showResidentes(List<Residente> residentes) {
+        residentesList.clear();
+        residentesList.addAll(residentes);
         // con esto la lista siempre estára actualizada cuando vuelva de un segundo plano.
         adapter.notifyDataSetChanged();
-
     }
 
     @Override
