+++ app/src/main/res/layout/activity_centros.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.CentrosListView">
+    tools:context=".view.centro.CentrosListView">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/topAppBar"
