+++ app/src/main/res/layout/activity_profesionales.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ProfesionalesListView">
+    tools:context=".view.profesional.ProfesionalesListView">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/topAppBar"
