+++ app/src/main/res/layout/activity_residentes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ResidentesListView">
+    tools:context=".view.residente.ResidentesListView">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/topAppBar"
